+++ modules/apps/portal/portal-async-advice/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Portal Async Advice
+Bundle-SymbolicName: com.liferay.portal.async.advice
+Bundle-Version: 1.0.0
